Reorganize code [WIP]
