+++ src/event/checked_event.rs
@@ -12,7 +12,7 @@ mod tag;
 mod visibility;
 
 pub use announcement::Announcement;
-pub(in crate::messages::event) use announcement_reaction::AnnouncementReaction;
+pub(in crate::event) use announcement_reaction::AnnouncementReaction;
 pub use conversation::Conversation;
 pub use id::Id;
 pub use notification::Notification;
