+++ src/event/checked_event/announcement.rs
