+++ src/event/checked_event/announcement_reaction.rs
