+++ src/event/checked_event/emoji.rs
