--- src/messages.rs
@@ -1,3 +1,0 @@
-mod event;
-
-pub use event::{CheckedEvent, DynEvent, Event, EventErr, EventKind, Id};
